tests/requirements.txt
flake8
coverage
django-simple-captcha
Pillow
pytz
